# src/main/resources/api.yaml
openapi: "3.0.1"
info:
  description: "TiAPI is a standalone service providing restful API to execute SQL statements on TiDB to users. It's an experimental project making TiDB more serviceless. With TiAPI, the real TiDB instances are transparent to end users. This frees users from management duties of database instances and offers more flexibility to the database providers of upgrading and other management things."
  version: "1.0.0"
  title: "TiAPI"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  /management/tidb:
    get:
      tags:
        - "management"
      summary: "list all backend TiDB instances"
      operationId: "listTiDB"
      responses:
        "200":
          description: "list all backend TiDB instances successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTiDBResponse"
    post:
      tags:
        - "management"
      summary: "add new backend TiDB instance"
      operationId: "addTiDB"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TiDBInstance'
        required: true
      responses:
        "201":
          description: "add a TiDB instance successfully"
...
components:
  schemas:
...
    TiDBInstance:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
        adminUsername:
          type: string
        adminPassword:
          type: string
    ListTiDBResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/TiDBInstance'
...